backend/requirements.txt
langchain==0.1.0
google-generativeai==0.3.0
fastapi==0.104.1
uvicorn==0.24.0
chromadb==0.4.18
beautifulsoup4==4.12.2
requests==2.31.0
python-dotenv==1.0.0
pydantic==2.5.2

//interaction_feed
# fastapi>=0.68.0
# uvicorn>=0.15.0
# requests>=2.26.0
textblob>=0.15.3
python-dateutil>=2.8.2
# pydantic>=1.8.2